// planning-idsdp-pps-frontend/src/app/main/shared/enum/user-group.enum.ts
export enum UserGroup {
    AGENCY_HEAD = 'AGENCY-HEAD',
    AGENCY_DESK = 'AGENCY-DESK',
    MINISTRY_HEAD = 'MINISTRY-HEAD',
    MINISTRY_DESK = 'MINISTRY-DESK',
    PLANNING_HEAD = 'PLANNING-HEAD',
    PLANNING_DESK = 'PLANNING-DESK',
    PLANNING_MINISTER = 'PLANNING-MINISTER',
    ECNEC = 'ECNEC-HEAD',
    ECNEC_OFFICER = 'ECNEC-OFFICER',
    ECNEC_DESK = 'ECNEC-DESK',
}